Modified escrow payment flow, renamed escrow_payment.dart to payment.dart, updated task details screen, and client model.
